pyproject.toml
[project]
name = "biblioteka"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "crispy-bootstrap4 (>=2025.6,<2026.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.3"
isort = "^6.0.1"

